requirements.txt
openai
pinecone-client
langchain
chainlit
